chore(deps): break repair-bay + unit-card-expanded cycles
Pattern 4 + direct re-route.

- src/components/gameplay/pages/repair/bay/RepairBayPage.RepairWorkspace.tsx
  Imported UnitRepairCard from RepairBayPage.sections (which only re-exports
  it from RepairBayPage.UnitRepairCard). Switched to direct import from
  RepairBayPage.UnitRepairCard.tsx so RepairWorkspace no longer depends on
  the orchestrator that imports it back.
- src/components/unit-card/UnitCardExpanded.tsx <-> UnitCardExpandedSections.tsx
  EquipmentEntry, CriticalSlotSummary, UnitCardExpandedProps moved to
  UnitCardExpanded.types.ts; both halves import the leaf and the
  orchestrator re-exports the types so unit-card/index.ts public API is
  preserved.

Verified: madge -> 0 circular deps in src/components/gameplay/pages/repair/bay
and src/components/unit-card.
